move ValidationElement helpers to their own file
