+++ src/components/Form/ValidationElement/helpers.test.js
@@ -1,6 +1,6 @@
-import { flattenObject, mergeError, triageErrors } from './ValidationElement'
+import { flattenObject, mergeError, triageErrors } from './helpers'
 
-describe('The ValidationElement component', () => {
+describe('The ValidationElement helpers', () => {
   it('can flatten an object', () => {
     const tests = [
       {
