+++ src/components/Form/ValidationElement/index.js
@@ -1,3 +1,4 @@
-import ValidationElement, { newGuid, flattenObject, mergeError, triageErrors } from './ValidationElement'
+import { newGuid } from './helpers'
+import ValidationElement from './ValidationElement'
 export default ValidationElement
-export { newGuid, flattenObject, mergeError, triageErrors }
+export { newGuid }
